requirements.txt
argcomplete>=0.8.8,<=1.1.1
docker-py>=1.7.2,<=1.8.1
